chang float to double

+++ src/paint/vector_operation.c
@@ -6,7 +6,7 @@
 /*   By: tomartin <tomartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/06 10:52:03 by tomartin          #+#    #+#             */
-/*   Updated: 2021/12/06 16:24:04 by tomartin         ###   ########.fr       */
+/*   Updated: 2021/12/08 19:20:02 by tomartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,15 +30,15 @@ t_vect	subtr_v(t_vect v_a, t_vect v_b)
 	return (answ);
 }
 
-float	mod_v(t_vect vect)
+double	mod_v(t_vect vect)
 {
-	float	answ;
+	double	answ;
 
 	answ = sqrtf(powf(vect.o.x, 2) + powf(vect.o.y, 2));
 	return (answ);
 }
 
-t_vect	proc_vect(t_vect v_a, float n)
+t_vect	proc_vect(t_vect v_a, double n)
 {
 	t_vect	answ;
 
@@ -47,7 +47,7 @@ t_vect	proc_vect(t_vect v_a, float n)
 	return (answ);
 }
 
-void	load_values_v(t_vect *load, float x, float y)
+void	load_values_v(t_vect *load, double x, double y)
 {
 	load->o.x = x;
 	load->o.y = y;
